[notification] setup mailhog for tests
In order to view the emails sent by the application, Mailhog had to be setup. Mailhog can be easily setup using Laravel Sail, which is made to run container environments. That being said, Sail was used to bootstrap a container environment in this application, following the Documentation at Laravel's website.